Thesis Stage1 wind up reference
I am not willing to do a more detailed wind-up here, so lets' just put the things together and check it when needed
```
Hi Tim, hi Fei, now we support custom plugins of ImageJ1/2 on jupyterlab. 【Figure 1 】is the Plugin compatibility check flowchart, 【Figure 2】 is the transparent warning feedback when user add their custom plugins, 【Figure 3】 is the lists of built-in ImageJ1 commands, ImageJ2 ops, and here's the 【Guideline documentations】 for users who would like to use their custom plugins of imageJ on jupyterlab, the reasons are stated in detail. 
```